File name is correctly adjusted.
When execute button is clicked, it makes a copy of the original file as "originalFileName_copied.extension" and make encrypted file "originalFileName_encrypted.extension". When decrypt button is clicked, "originalFileName_decrypted.extension" is created. Both copied and decrypted files are same as the original file, while encrypted file couldn't be opened on Windows.